Rename configuration file
