+++ test/acceptance/cartocss-transformations-test.js
@@ -4,7 +4,7 @@ require('../support/test-helper');
 
 var assert = require('../support/assert');
 var TestClient = require('../support/test-client');
-const environment = require('../support/environment');
+const config = require('../support/config');
 
 var IMAGE_EQUALS_TOLERANCE_PER_MIL = 20;
 
@@ -201,7 +201,7 @@ describe('cartocss transformation', function () {
                 mapnik: {
                     grainstore: Object.assign(
                         {},
-                        environment.renderer.mapnik.grainstore,
+                        config.renderer.mapnik.grainstore,
                         { mapnik_version: scenario.mapnikVersion }
                     )
                 }
@@ -242,7 +242,7 @@ describe('cartocss transformation', function () {
                 mapnik: {
                     grainstore: Object.assign(
                         {},
-                        environment.renderer.mapnik.grainstore,
+                        config.renderer.mapnik.grainstore,
                         { mapnik_version: scenario.mapnikVersion }
                     )
                 }
