+++ test/acceptance/limits-test.js
@@ -6,7 +6,7 @@ var fs = require('fs');
 
 var assert = require('../support/assert');
 var TestClient = require('../support/test-client');
-const environment = require('../support/environment');
+const config = require('../support/config');
 
 describe('render limits', function () {
     function onTileErrorStrategyPass (err, tile, headers, stats, format, callback) {
@@ -29,7 +29,7 @@ describe('render limits', function () {
 
     var OVERRIDE_OPTIONS = {
         mapnik: {
-            mapnik: Object.assign({}, environment.renderer.mapnik.mapnik, { limits: LIMITS_CONFIG })
+            mapnik: Object.assign({}, config.renderer.mapnik.mapnik, { limits: LIMITS_CONFIG })
         }
     };
