+++ test/acceptance/mapnik-render-by-format-quality-test.js
@@ -6,7 +6,7 @@ var assert = require('../support/assert');
 var fs = require('fs');
 var TestClient = require('../support/test-client');
 const path = require('path');
-const environment = require('../support/environment');
+const config = require('../support/config');
 
 var IMAGE_EQUALS_TOLERANCE_PER_MIL = 85;
 
@@ -15,7 +15,7 @@ describe('mapnik_render_by_format_quality', function () {
     before(function (done) {
         testClient = new TestClient(layergroup, {
             mapnik: {
-                grainstore: Object.assign({ mapnik_tile_format: 'png8:m=h' }, environment.renderer.mapnik.grainstore)
+                grainstore: Object.assign({ mapnik_tile_format: 'png8:m=h' }, config.renderer.mapnik.grainstore)
             }
         });
         var testPngFilesDir = path.join(__dirname, '/../results/png');
